Create fine-tune

Creates a job that fine-tunes a specified model from a given dataset. See this page for details.

Usage

create_fine_tune(
  training_file,
  validation_file = NULL,
  model = c("curie", "ada", "babbage", "davinci"),
  n_epochs = 4,
  batch_size = NULL,
  learning_rate_multiplier = NULL,
  prompt_loss_weight = 0.1,
  compute_classification_metrics = FALSE,
  classification_n_classes = NULL,
  classification_positive_class = NULL,
  classification_betas = NULL,
  suffix = NULL,
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL
)

Arguments

training_file

required; a length one character vector. The ID of an uploaded file that contains training data. See upload_file() for how to upload a file. Your dataset must be formatted as a JSONL file, where each training example is a JSON object with the keys "prompt" and "completion". Additionally, you must upload your file with the purpose fine-tune. See the fine-tuning guide for more details.

validation_file

optional; defaults to NULL; a length one character vector. The ID of an uploaded file that contains validation data. If you provide this file, the data is used to generate validation metrics periodically during fine-tuning. These metrics can be viewed in the fine-tuning results file. Your train and validation data should be mutually exclusive. Your dataset must be formatted as a JSONL file, where each validation example is a JSON object with the keys "prompt" and "completion". Additionally, you must upload your file with the purpose fine-tune. See the fine-tuning guide for more details.

model

model required; defaults to "curie"; a length one character vector, one among "ada", "babbage", "curie", and "davinci". The name of the base model to fine-tune. To learn more about these models, see the Engines documentation.

n_epochs

required; defaults to 4; a length one numeric vector with the integer value greater than 0. The number of epochs to train the model for. An epoch refers to one full cycle through the training dataset.

batch_size

optional; defaults to NULL; a length one numeric vector with the integer value greater than 0. The batch size to use for training. The batch size is the number of training examples used to train a single forward and backward pass. By default, the batch size will be dynamically configured to be ~0.2% of the number of examples in the training set, capped at 256 - in general, we've found that larger batch sizes tend to work better for larger datasets.

learning_rate_multiplier

optional; defaults to NULL; a length one numeric vector with the value greater than 0. The learning rate multiplier to use for training. The fine-tuning learning rate is the original learning rate used for pretraining multiplied by this value. By default, the learning rate multiplier is the 0.05, 0.1, or 0.2 depending on final batch_size (larger learning rates tend to perform better with larger batch sizes). We recommend experimenting with values in the range 0.02 to 0.2 to see what produces the best results.

prompt_loss_weight

required; defaults to 0.1; a length one numeric vector. The weight to use for loss on the prompt tokens. This controls how much the model tries to learn to generate the prompt (as compared to the completion which always has a weight of 1.0), and can add a stabilizing effect to training when completions are short. If prompts are extremely long (relative to completions), it may make sense to reduce this weight so as to avoid over-prioritizing learning the prompt.

compute_classification_metrics

required; defaults to FLASE; a length one logical vector. If set, we calculate classification-specific metrics such as accuracy and F-1 score using the validation set at the end of every epoch. These metrics can be viewed in the results file. In order to compute classification metrics, you must provide a validation_file. Additionally, you must specify classification_n_classes for multiclass classification or classification_positive_class for binary classification.

classification_n_classes

optional; defaults to NULL; a length one numeric vector with the value greater than 0. The number of classes in a classification task. This parameter is required for multiclass classification.

classification_positive_class

optional; defaults to NULL; a length one character vector. The positive class in binary classification. This parameter is needed to generate precision, recall, and F1 metrics when doing binary classification.

classification_betas

optional; defaults to NULL; a list elements of which are numeric values greater than 0. If this is provided, we calculate F-beta scores at the specified beta values. The F-beta score is a generalization of F-1 score. This is only used for binary classification. With a beta of 1 (i.e. the F-1 score), precision and recall are given the same weight. A larger beta score puts more weight on recall and less on precision. A smaller beta score puts more weight on precision and less on recall.

suffix

optional; defaults to NULL; a length one character vector. A string of up to 40 characters that will be added to your fine-tuned model name. For example, a suffix of "custom-model-name" would produce a model name like ada:ft-your-org:custom-model-name-2022-02-15-04-21-04.

openai_api_key

required; defaults to Sys.getenv("OPENAI_API_KEY") (i.e., the value is retrieved from the .Renviron file); a length one character vector. Specifies OpenAI API key.

openai_organization

optional; defaults to NULL; a length one character vector. Specifies OpenAI organization.

Value

Returns a list, elements of which contain information about the fine-tune.

Details

Manage fine-tuning jobs to tailor a model to your specific training data. Creates a job that fine-tunes a specified model from a given dataset. Response includes details of the enqueued job including job status and the name of the fine-tuned models once complete. Related guide: Fine-tune models.

See also

Examples

if (FALSE) {
training_file <- system.file(
    "extdata", "sport_prepared_train.jsonl", package = "openai"
)
validation_file <- system.file(
    "extdata", "sport_prepared_train.jsonl", package = "openai"
)

training_info <- upload_file(training_file, "fine-tune")
validation_info <- upload_file(validation_file, "fine-tune")

info <- create_fine_tune(
    training_file = training_info$id,
    validation_file = validation_info$id,
    model = "ada",
    compute_classification_metrics = TRUE,
    classification_positive_class = " baseball" # Mind space in front
)
}