Upload file

Uploads a file that will be used for various purposes. The size of the storage is limited to 1 Gb. See this page for details.

Usage

upload_file(
  file,
  purpose = c("search", "answers", "classifications", "fine-tune"),
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL
)

Arguments

file

required; a length one character vector.

purpose

required; defaults to "fine-tune"; a length one character vector equals to"fine-tune".

openai_api_key

required; defaults to Sys.getenv("OPENAI_API_KEY") (i.e., the value is retrieved from the .Renviron file); a length one character vector. Specifies OpenAI API key.

openai_organization

optional; defaults to NULL; a length one character vector. Specifies OpenAI organization.

Value

Returns a list, elements of which contains ID of the uploaded file and other supplementary information.

Details

For arguments description please refer to the official documentation.

See also

Examples

if (FALSE) { # \dontrun{
file <- system.file("extdata", "classification-file.jsonl", package = "openai")
upload_file(file = file, purpose = "classification")
} # }